% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_footer.R
\name{add_footer}
\alias{add_footer}
\title{Add a formatted footer to a flextable}
\usage{
add_footer(ft, text = "Sample footnote", fontsize = 10,
  fontname = "Times New Roman")
}
\arguments{
\item{ft}{A flextable object}

\item{text}{The footer text}

\item{fontsize}{The fontsize}

\item{fontname}{The fontname}
}
\value{
A flextable object
}
\description{
Add a formatted footer to a flextable
}
