% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tables.R
\name{add_table}
\alias{add_table}
\title{Add tables and titles to a word document}
\usage{
add_table(.docx, tables, titles = NA)
}
\arguments{
\item{.docx}{An officer object}

\item{tables}{A list of tables.}

\item{titles}{A list of titles.  Must be either NA or the length of tables.  If NA, tables will be
titles 'Table 1', 'Table 2', ...}
}
\description{
Extends the functionality of the officer package by easing the creation of word documents.  Tables
will be added sequentually.  A page break will be inserted after every table except the last.
}
