% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_total_column.R
\name{add_total_column}
\alias{add_total_column}
\title{Add a total column to a tbl}
\usage{
add_total_column(.data, name = "Total")
}
\arguments{
\item{.data}{A tbl.}

\item{name}{The output column name. If omitted, it will be Total.}
}
\value{
An object of the same class as .data
}
\description{
Sum up all numeric columns in a data frame
}
\examples{
df <- data.frame(id = LETTERS[1:3], x1 = 1:3, x2 = 4:6, x3 = 7:9)
df
df \%>\% add_total_column()
}
