% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_unique.R
\name{all_unique}
\alias{all_unique}
\title{Check if all values are unique}
\usage{
all_unique(vec)
}
\arguments{
\item{vec}{A vector or list.}
}
\description{
Checking if all values are unique can be a good step in QC to ensure entries are not duplicated.  If passed a
list, the input is first flattened (which may result in type casting).
}
\examples{
all_unique(c(2,1,3))
all_unique(c(1,1,2))
all_unique(list(1, "1", 2))
}
