% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{Calculate age}
\usage{
calc_age(dob, ref_date = Sys.Date())
}
\arguments{
\item{dob}{Starting date or datetime (usually date of birth).}

\item{ref_date}{Ending date or datetime}
}
\description{
This function is just a convenient wrapper to calculate age using lubridate.
}
\examples{
calc_age(dob = as.Date("1974-03-14"))

birth_days <- sample(seq(as.Date("1999/01/01"), as.Date("2000/01/01"), by = "day"), 5)
follow_ups <- data.frame(dob = birth_days, last_checkup = Sys.Date())

follow_ups \%>\%
  dplyr::mutate(age_last_fu = calc_age(dob, last_checkup))
}
