% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_rows.R
\name{coalesce_rows}
\alias{coalesce_rows}
\title{Coalesce columns of a data frame}
\usage{
coalesce_rows(.data)
}
\arguments{
\item{.data}{A tbl.}
}
\description{
Often helpful when data has duplicate rows, but column data is split between entries.
}
\examples{
df <- tibble::tribble(
  ~V1, ~V2, ~V3, ~V4,
  "A", 1,NA,2,
  "B", 3,4,3,
  "C", 5,6,3,
  "A", 2,3,NA
)

df \%>\%
  dplyr::group_by(V1) \%>\%
  coalesce_rows()
}
