% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countpct.R
\name{countpct}
\alias{countpct}
\title{Summarize two columns as N (\%)}
\usage{
countpct(.data, n, perc, name = "N", accuracy = 0.01, remove = TRUE)
}
\arguments{
\item{.data}{A tbl.}

\item{n}{Column of values.}

\item{perc}{Column of percentages (as decimals).}

\item{name}{Name of new column.}

\item{accuracy}{Number to round to, NULL for automatic guess.}

\item{remove}{If TRUE, remove input columns from output data frame.}
}
\description{
Provided two columns, create a new column representing a number and percentage
}
\examples{
iris \%>\%
  dplyr::group_by(Species) \%>\%
  dplyr::summarise(n = dplyr::n()) \%>\%
  dplyr::ungroup() \%>\%
  dplyr::mutate(freq = prop.table(n)) \%>\%
  countpct(n, freq)
}
