% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{days.abb}
\alias{days.abb}
\title{Three-letter abbreviations for days of week}
\format{A character vector of length 7.}
\usage{
days.abb
}
\description{
Three-letter abbreviations for days of the week starting with Sunday
}
\examples{
days.abb
}
\keyword{datasets}
