% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_flextable.R
\name{format_flextable}
\alias{format_flextable}
\title{Standard options used to format a flextable}
\usage{
format_flextable(
  data,
  center_cols = NULL,
  left_cols = NULL,
  right_cols = NULL,
  header1 = NULL,
  header2 = NULL,
  bold_header = FALSE,
  center_num = TRUE,
  fontsize = 10,
  fontname = "Times New Roman",
  pbold = NA,
  psig = 0.05
)
}
\arguments{
\item{data}{A `data frame` or `flextable` object}

\item{center_cols, left_cols, right_cols}{numeric vector of columns to center, left or right align}

\item{header1, header2}{Named list of values and colwidths.  header1 will replace the original header.  header2 will add a
second header on top.}

\item{bold_header}{Bold the header column(s)?}

\item{center_num}{Center numeric columns?}

\item{fontsize, }{fontsize}

\item{fontname}{fontname}

\item{pbold}{(String) Name of column to bold as p-values}

\item{psig}{A numeric threshold to bold p-values#' @param psig A numeric threshold to bold p-values}
}
\description{
Standard options used to format a flextable
}
