% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_modelsum.R
\name{format_modelsum}
\alias{format_modelsum}
\title{Generate a formatted flextable from a modelsum object}
\usage{
format_modelsum(tbl)
}
\arguments{
\item{tbl}{A modlesum object}
}
\value{
A flextable object
}
\description{
Generate a formatted flextable from a modelsum object
}
