% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{Format a column of p-values}
\usage{
format_p(pval, na_string = "-", empty_string = "-", sig = NA)
}
\arguments{
\item{pval}{A vector of p values}

\item{na_string}{Value to return if pval is NA}

\item{empty_string}{Value to return if pval == ""}

\item{sig}{Level of signficance.  If formatted p-value = significance level, an extra digit will be added}
}
\description{
Format a column of p-values
}
