% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tableby.R
\name{format_tableby}
\alias{format_tableby}
\title{Generate a formatted flextable from a tableby object}
\usage{
format_tableby(tbl, bold_var = T, bold_p = T, pvals = "p value", ...)
}
\arguments{
\item{tbl}{A tableby object}

\item{bold_var}{Bold variable names in the first column?}

\item{bold_p}{(Logical) If TRUE, p values less than 0.05 will be bolded.}

\item{pvals}{A character vector of p values to format}

\item{...}{Additional arguments to pass to format_flextable}
}
\description{
Generate a formatted flextable from a tableby object
}
\examples{
\dontrun{
tableby(arm ~ ., data = mockstudy, digits = 2) \%>\% format_tableby()
}

}
