% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_row_header.R
\name{in_row_headers}
\alias{in_row_headers}
\title{Add header rows within a column for values that repeat.}
\usage{
in_row_headers(.data, col)
}
\arguments{
\item{.data}{A tbl.}

\item{col}{Column to add subheaders to.}
}
\description{
Often helpful when wanting to create sub-groupings within a column in a final output.
}
\examples{
df <- tibble::tribble(
  ~Label, ~value,
  "A", 1.2,
  "A", 1.3,
  "B", 5.0,
  "C", 4.3,
  "C", 4.1,
  "C", 5.2,
  "D", 6
)
df \%>\% in_row_headers(Label)
}
