% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitthis.R
\name{knitthis}
\alias{knitthis}
\title{Knit the Rmd file that is currently open}
\usage{
knitthis(output_dir = "documents", ...)
}
\arguments{
\item{output_dir}{The output directory, as used in \code{rmarkdown::render}}

\item{...}{Additional arguments to pass to \code{rmarkdown::render}}
}
\description{
This is a convenient alternative to using the knit button
when you want to specify additional arguments such as output directory.
}
\examples{
\dontrun{
# This command will knit the Rmd file that is currently open into the directory R.
knitthis("R")

# Assuming the open file is data_prep.Rmd, this would be equivalent to running:
rmarkdown::render(".../data_prep.Rmd", output_dir = ".../R")

}

}
