% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_df.R
\name{label_df}
\alias{label_df}
\title{Create a data frame with columns for variables and variable labels}
\usage{
label_df(.data)
}
\arguments{
\item{.data}{A tbl.}
}
\description{
Often helpful when wanting to replace variable names with labels in final output.
}
