% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsum_cox.R
\name{modelsum_cox}
\alias{modelsum_cox}
\title{Univariable Cox regression Table}
\usage{
modelsum_cox(event, time, data = NA, pval = TRUE)
}
\arguments{
\item{event}{Name of event column}

\item{time}{Name of time column}

\item{data}{A tbl.}

\item{pval}{Include p-value?}
}
\description{
Univariable Cox regression Table
}
