% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsum_km.R
\name{modelsum_km}
\alias{modelsum_km}
\title{Kaplan-Meier Table}
\usage{
modelsum_km(event, time, data = NA, times = 1:3,
  freedom.from = "death", pval = T)
}
\arguments{
\item{event}{Name of event column}

\item{time}{Name of time column}

\item{data}{A tbl.}

\item{times}{Numeric vector of time points to analyze}

\item{freedom.from}{Character string to be used in table header}

\item{pval}{Include p value?}
}
\description{
Kaplan-Meier Table
}
