% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_column.R
\name{move_to_last}
\alias{move_to_last}
\alias{move_to_first}
\title{Move a column to the beginning or end of a data frame}
\usage{
move_to_last(.data, col)

move_to_first(.data, col)
}
\arguments{
\item{.data}{A tbl.}

\item{col}{Name of column to move (unquoted)}
}
\value{
An object of the same class as .data
}
\description{
Move a column to the beginning or end of a data frame
}
\examples{
head(iris)
head(iris) \%>\% move_to_first(Species)
head(iris) \%>\% move_to_last(Sepal.Length)
}
