% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrclwhip.R
\docType{package}
\name{mRclwhip}
\alias{mRclwhip}
\title{Whippin' Data Into Shape at Mayo Clinic}
\description{
Quickly manipulate data and prepare publication-ready tables.
}
\details{
\code{mRclwhip} package
}
