% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Create a new R project and accompanying folders}
\usage{
new_project(project_dir = "/projects/bsi/fl/studies",
  project_name = "sXXXXXX.new.study", launch = TRUE,
  new_session = TRUE, rproj = T, skeleton = c("rawdata", "documents",
  "figures", "data"))
}
\arguments{
\item{project_dir}{Name of directory to put project folder in}

\item{project_name}{Name of project folder}

\item{launch}{Launch new project}

\item{new_session}{If launching project, launch in a new session?}

\item{rproj}{Create a .Rproj file?}

\item{skeleton}{Folders to create inside project directory}
}
\description{
Create a new project directory and accompanying folders
}
\examples{
# new_project(s123456.PI.Project)
# The above command would create a the following folder structure:
#  |--- /projects/bsi/fl/studies
#      |--- s123456.PI.Project
#          |--- rpgm
#              |--- rpgm.Rproj
#              |--- rawdata/
#              |--- documents/
#              |--- figures/
#              |--- data/
}
