% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_na.R
\name{non_na}
\alias{non_na}
\title{Return indexes of values that are (or are not) NA}
\usage{
non_na(x, negate = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{negate}{If `TRUE`, return indexes of NA values.}
}
\description{
Numeric indexes of non-NA values in a vector
}
\examples{
non_na(c(1, 2, NA, 3, NA))
non_na(c(1, 2, NA, 3, NA), negate = TRUE)
}
