% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrange.R
\name{nrange}
\alias{nrange}
\title{Summarize three columns as N (low, high)}
\usage{
nrange(
  .data,
  n,
  low,
  high,
  accuracy = 0.01,
  name = "Estimate",
  formatter = c("comma", "percent", "number"),
  remove = TRUE,
  ...
)
}
\arguments{
\item{.data}{A tbl.}

\item{n, low, high}{Column names to summarize}

\item{accuracy}{Number to round to, NULL for automatic guess.}

\item{name}{Name of new column.}

\item{formatter}{scales function used to format text}

\item{remove}{If TRUE, remove input columns from output data frame.}
}
\value{
An object of the same class as .data.
}
\description{
Often useful for summarizing an association metric and it's confidence interval.
}
\examples{
df <- tibble::tribble(
  ~Group, ~n, ~low, ~high,
  "a", 1.323453, 1.154, 1.44,
  "b", .801, .741, .9891
)

df \%>\% nrange(n, low, high)
}
