% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_change.R
\name{pct_change}
\alias{pct_change}
\title{Calculate the percent change between two values or vectors of values}
\usage{
pct_change(start, final, perc = FALSE)
}
\arguments{
\item{start}{A value or vector of values}

\item{final}{A value or vector of values}

\item{perc}{A logical indicating whether or not to format the value with a percent symbol}
}
\description{
Calculate the percent change between two values or vectors of values
}
\examples{
tibble(
 start = seq(0.1, 1.1, 0.1),
 final = sample(seq(0.1, 1, 0.1), 11, replace = T)
 ) \%>\%
mutate(pct_diff = pct_change(start, final))
}
