% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_to_numeric.R
\name{percent_to_numeric}
\alias{percent_to_numeric}
\title{Convert a formatted percent to a numeric}
\usage{
percent_to_numeric(perc)
}
\arguments{
\item{perc}{A vector of strings}
}
\description{
Convert a vector of percents (strings) to a numeric vector.
}
\examples{
percent_to_numeric(c("32.4\%", "1.2\%"))
}
