% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{polish}
\alias{polish}
\title{Tidy and put the finishing touches on a model output}
\usage{
polish(
  x,
  .labels = NULL,
  .conf_int = TRUE,
  .flextable = TRUE,
  .header1 = NULL,
  .header2 = NULL,
  ...
)
}
\arguments{
\item{x}{A model object to be converted into a tibble::tibble()}

\item{.labels}{A named list of variable labels. If NULL, current labels will be used}

\item{.conf_int}{Logical to return a 95% confidence interval}

\item{.flextable}{Logical to return a formatted flextable object if TRUE or a list of a dataframe and row numbers to use in flextable::padding()}

\item{.header1}{See ??mRclwhip::format_flextable for details}

\item{.header2}{See ??mRclwhip::format_flextable for details}

\item{...}{additional arguments passed to broom::tidy()}
}
\value{
A flextable::flextable() or a list with a tidy tibble::tibble() a vector of row numbers
}
\description{
Wrapper around broom::tidy() with some additional finishing touches!
}
\examples{
df <- tibble::tribble(
  ~id, ~age, ~mile_time,
  "a",  23,   7,
  "b",  33,   6.5,
  "c",  18,   7.25,
  "d",  27,   6.8,
)

lm_res <- lm(mile_time ~ age, data = df)

lm_res \%>\% polish(.header1 = list(values = c("", "Estimate (95\% CI)")))
}
