% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_label.R
\name{sort_by_label}
\alias{sort_by_label}
\title{Sort the columns of a data frame by it's labels (alphabetically)}
\usage{
sort_by_label(.data)
}
\arguments{
\item{.data}{A tbl}
}
\value{
An object of the same class as .data
}
\description{
Helpful when using labels in the final output
}
