% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_nrange.R
\name{split_nrange}
\alias{split_nrange}
\title{Split character vector of N (low, high) into three columns}
\usage{
split_nrange(.data, col, remove = TRUE)
}
\arguments{
\item{.data}{A tbl.}

\item{col}{Column to separate}

\item{remove}{If TRUE, remove input columns from output data frame.}
}
\value{
An object of the same class as .data.
}
\description{
Split character vector of N (low, high) into three columns
}
\examples{
tmp <- data.frame(
  obs = "A",
  val = "1224.11 (119.3214, 134.21)",
  stringsAsFactors = FALSE
)

tmp \%>\%
  split_nrange(val) \%>\%
  nrange(n, low, high)
}
