% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_firstup.R
\name{str_to_firstup}
\alias{str_to_firstup}
\title{Capitalize the first character of a string}
\usage{
str_to_firstup(string)
}
\arguments{
\item{string}{String to modify}
}
\description{
Capitalize the first character of a string without altering case of any other character.
}
\examples{
str_to_firstup("age at death")
str_to_firstup("baseline BMI")
c("age at death", "baseline BMI") \%>\% str_to_firstup()
}
