% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfill_vec.R
\name{unfill_vec}
\alias{unfill_vec}
\title{Unfill a vector}
\usage{
unfill_vec(x)
}
\arguments{
\item{x}{a vector of values}
}
\description{
The inverse of dplyr::fill.
Useful for formatting data frames for final output, where one often wants to avoid repeating values.
Often used in conjunction with \code{in_row_headers()}.
}
\examples{
move_to_first(iris, Species) \%>\%
  in_row_headers(Species) \%>\%
  dplyr::mutate(Species = unfill_vec(Species))
}
\seealso{
\code{in_row_headers()}
}
