% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_to_regex.R
\name{vec_to_regex}
\alias{vec_to_regex}
\title{Convert a vector to a regular expression}
\usage{
vec_to_regex(..., sep = "|", wrap = c("(", ")"))
}
\arguments{
\item{...}{Vectors to collapse into the regex.  Duplicated elements are removed.}

\item{sep}{Character used to separate elements of ...}

\item{wrap}{2-element vector of strings to wrap elements of ... in.}
}
\description{
Often helpful when constructing a regular expression based on a vector of values
}
\examples{
vec_to_regex(month.abb)

vec_to_regex(letters[1:5], sep = "", wrap = c("[", "]+"))
}
