% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{polish}
\alias{polish}
\title{Tidy and put the finishing touches on a model output}
\usage{
polish(
  x,
  .labels = NULL,
  .conf_int = TRUE,
  .flextable = TRUE,
  .header1 = NULL,
  .header2 = NULL,
  ...
)
}
\arguments{
\item{x}{A model object to be converted into a \code{tibble::tibble()}}

\item{.labels}{A named list of variable labels. If NULL, current labels will be used}

\item{.conf_int}{Logical to return a 95 percent confidence interval}

\item{.flextable}{Logical to return a formatted flextable object if TRUE or a list of a dataframe and row numbers to use in \code{flextable::padding()}}

\item{.header1}{See \code{??mRclwhip::format_flextable} for details}

\item{.header2}{See \code{??mRclwhip::format_flextable} for details}

\item{...}{additional arguments passed to \code{broom::tidy()}}
}
\description{
Wrapper around \code{broom::tidy()} with some additional finishing touches!
}
\examples{
df1 <- tibble::tibble(
  id = letters,
  age = sample(seq(18, 39, 1), 26, replace = T),
  mile_time = sample(seq(6, 7.5, 0.1), 26, replace = T),
  gender = sample(c("Male", "Female"), 26, replace = T),
  country = sample(c("USA", "Canada", "Africa", "England"), 26, replace = T)
)

lm_res <- lm(mile_time ~ age + gender + country, data = df1)

lm_res \%>\%
  polish(
    .labels = c(age = "Age", gender = "Gender", country = "Country of origin"),
    .header1 = list(values = c("", "Estimate (95\% CI)"))
    )
}
